clean up games forms
