+++ hgapp/games/urls.py
@@ -29,6 +29,10 @@ urlpatterns = [
     # ex: .com/games/edit/g/112
     url(r'^edit/g/(?P<game_id>[\d]+)/$', views.edit_game, name='games_game_edit'),
 
+    # ex: .com/games/edit-completed/g/112/p/9520+323+412/
+    url(r'^edit-completed/g/(?P<game_id>[\d]+)/p/(?P<players>[\d\+]*)/$',
+        views.edit_completed,
+        name='games_finalize_add_attendance'),
 
     # ex: .com/games/create/game/ex/cell/1
     url(r'^create/game/ex/cell/(?P<cell_id>[\d]+)/$',
@@ -39,6 +43,11 @@ urlpatterns = [
     url(r'^create/game/ex/cell/(?P<cell_id>[\d]+)/gm/(?P<gm_user_id>[\d]+)/p/(?P<players>[\d\+]+)/$',
         views.finalize_create_ex_game_for_cell,
         name='games_edit_ex_game_add_players'),
+
+    # ex: .com/games/edit/attendance/g/112/
+    url(r'^add/attendance/g/(?P<game_id>[\d]+)/$',
+        views.add_attendance,
+        name='games_add_attendance'),
 
     # ex: .com/games/start/g/112
     # ex: .com/games/start/g/112/char_error=21,1,3
